Move consistency check yaml from api to core module (#21364)
